Add context parameter, no functionality yet
The underlying slh functions now take ctx as a parameter and check it's length according to FIPS 205. But in this commit, the array is not yet used in any way. Traits were updated so that existing code and tests still compile and the empty context string is passed in unless the _with_context_ versions are called.